refactor: DrawCommand 패키지 변경 및 의존성 제거
