+++ src/test/java/blackjack/view/DrawCommandTest.java
@@ -1,8 +1,9 @@
-package blackjack.domain.game;
+package blackjack.view;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import blackjack.view.input.DrawCommand;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
